+++ src/common/value-limits.js
@@ -1,41 +1,29 @@
 // Upper limits for string length and number values.
 
-const territoryLimit = 8;								// Territory name.
-const countryLimit = 32;								// Country name.
-const stateLimit = 32;									// State name.
-const cityLimit = 32;									// City name.
-const dealSizeLimit = 16;								// Deal Size name.
-const productLineLimit = 16;							// Product Line name.
-const statusLimit = 16;									// Order Status name.
-const productCodeLimit = 32;							// Product Code.
-const currencyLimit = 99999;							// Maximum currency.
-const customerLimit = 128;								// Customer name.
-const phoneNumberLimit = 20;							// Phone number.
-const contactLimit = 32;								// Individual contact name.
-const addressLimit = 128;								// Address line.
-const postalCodeLimit = 16;								// Postal Code.
-const orderNumberLimit = 32767;							// Maximum order number.
-const orderItemLimit = 255;								// Maximum items per order.
-const orderQuantityLimit = 255;							// Maximum quantity.
-
-
-module.exports =
+function defineValueLimits()
 {
-	territory: territoryLimit,
-	country: countryLimit,
-	state: stateLimit,
-	city: cityLimit,
-	dealSize: dealSizeLimit,
-	productLine: productLineLimit,
-	orderStatus: statusLimit,
-	productCode: productCodeLimit,
-	currency: currencyLimit,
-	customer: customerLimit,
-	phoneNumber: phoneNumberLimit,
-	contact: contactLimit,
-	address: addressLimit,
-	postalCode: postalCodeLimit,
-	orderNumber: orderNumberLimit,
-	orderItem: orderItemLimit,
-	orderQuantity: orderQuantityLimit
-};
+	var defineRes = {};
+	
+	defineRes["territory"] = 8;								// Territory name.
+	defineRes["country"] = 32;								// Country name.
+	defineRes["state"] = 32;								// State name.
+	defineRes["city"] = 32;									// City name.
+	defineRes["dealSize"] = 16;								// Deal Size name.
+	defineRes["productLine"] = 16;							// Product Line name.
+	defineRes["orderStatus"] = 16;							// Order Status name.
+	defineRes["productCode"] = 32;							// Product Code.
+	defineRes["currency"] = 99999;							// Maximum currency.
+	defineRes["customer"] = 128;							// Customer name.
+	defineRes["phoneNumber"] = 20;							// Phone number.
+	defineRes["contact"] = 32;								// Individual contact name.
+	defineRes["address"] = 128;								// Address line.
+	defineRes["postalCode"] = 16;							// Postal Code.
+	defineRes["orderNumber"] = 32767;						// Maximum order number.
+	defineRes["orderItem"] = 255;							// Maximum items per order.
+	defineRes["orderQuantity"] = 255;						// Maximum quantity.
+	
+	return defineRes;
+}
+
+
+module.exports = defineValueLimits();